EL-40 : Support for `on_cargo_pod_finished_ascending` * Added full support for event `on_cargo_pod_finished_ascending` which now contains the event data `event`, `launched_by_rocket`, `player`, and `cargo_pod`. `player` and `cargo_pod` contain the details of if a player is present in the pod and what the pod's inventory is. * Added full support for event `on_rocket_launch_ordered` which now contains the event data `event`, `player`, and `silo`. `player` and `silo` contain if a player was riding the rocket and the details of which silo the rocket was launched from. * Added full support for event `on_rocket_launched` which now contains the event data `event`, and `silo`. `silo` contains the details of which silo the rocket was launched from. * Addressed assigning unused `index` variable within `has_value` helper. * Separated out event handlers into multiple files, and added registration of the handlers to the `events` table space, simplifying identification of events and organization of the code.
